+++ aws-datastore/src/main/java/com/amplifyframework/datastore/syncengine/MutationProcessor.java
@@ -142,15 +142,15 @@ final class MutationProcessor {
             .andThen(publishToNetwork(mutationOutboxItem)
                 .map(modelWithMetadata -> ensureModelHasSchema(mutationOutboxItem, modelWithMetadata))
                 .flatMapCompletable(modelWithMetadata ->
-                    // Once the server knows about it, it's safe to remove from the outbox.
-                    // This is done before merging, because the merger will refuse to merge
-                    // if there are outstanding mutations in the outbox.
-                    mutationOutbox.remove(mutationOutboxItem.getMutationId())
-                        .andThen(merger.merge(modelWithMetadata))
-                        .doOnComplete(() -> {
-                            String modelName = mutationOutboxItem.getModelSchema().getName();
-                            announceMutationProcessed(modelName, modelWithMetadata);
-                        })
+                            // Once the server knows about it, it's safe to remove from the outbox.
+                            // This is done before merging, because the merger will refuse to merge
+                            // if there are outstanding mutations in the outbox.
+                            mutationOutbox.remove(mutationOutboxItem.getMutationId())
+                                .andThen(merger.merge(modelWithMetadata))
+                                .doOnComplete(() -> {
+                                    String modelName = mutationOutboxItem.getModelSchema().getName();
+                                    announceMutationProcessed(modelName, modelWithMetadata);
+                                })
                 )
             )
             .doOnComplete(() -> {
@@ -193,12 +193,12 @@ final class MutationProcessor {
     ) {
         final SerializedModel originalModel = (SerializedModel) modelWithMetadata.getModel();
         final SerializedModel newModel = SerializedModel.builder()
-            .serializedData(SerializedModel.parseSerializedData(
-                    originalModel.getSerializedData(),
-                    modelSchema.getName(),
-                    schemaRegistry
-            ))
-            .modelSchema(modelSchema)
+                .modelSchema(modelSchema)
+                .serializedData(SerializedModel.parseSerializedData(
+                        originalModel.getSerializedData(),
+                        modelSchema.getName(),
+                        schemaRegistry
+                ))
             .build();
         return new ModelWithMetadata<>(newModel, modelWithMetadata.getSyncMetadata());
     }