+++ aws-storage-s3/src/main/java/com/amplifyframework/storage/s3/service/AWSS3StorageService.java
@@ -15,26 +15,17 @@
 
 package com.amplifyframework.storage.s3.service;
 
-import android.app.Notification;
-import android.app.NotificationChannel;
-import android.app.NotificationManager;
 import android.content.Context;
-import android.content.Intent;
-import android.os.Build;
 import androidx.annotation.NonNull;
-import androidx.annotation.RequiresApi;
-import androidx.core.app.NotificationCompat;
 
 import com.amplifyframework.storage.StorageException;
 import com.amplifyframework.storage.StorageItem;
 import com.amplifyframework.storage.s3.CognitoAuthProvider;
-import com.amplifyframework.storage.s3.R;
 import com.amplifyframework.storage.s3.utils.S3Keys;
 import com.amplifyframework.util.UserAgent;
 
 import com.amazonaws.ClientConfiguration;
 import com.amazonaws.mobileconnectors.s3.transferutility.TransferObserver;
-import com.amazonaws.mobileconnectors.s3.transferutility.TransferService;
 import com.amazonaws.mobileconnectors.s3.transferutility.TransferUtility;
 import com.amazonaws.mobileconnectors.s3.transferutility.UploadOptions;
 import com.amazonaws.regions.Region;
@@ -64,8 +55,6 @@ public final class AWSS3StorageService implements StorageService {
     private final TransferUtility transferUtility;
     private final AmazonS3Client client;
     private final CognitoAuthProvider cognitoAuthProvider;
-
-    private boolean transferUtilityServiceStarted = false;
 
     /**
      * Constructs a new AWSS3StorageService.
@@ -136,7 +125,7 @@ public final class AWSS3StorageService implements StorageService {
             @NonNull String serviceKey,
             @NonNull File file
     ) {
-        startServiceIfNotAlreadyStarted();
+        startTransferService();
         return transferUtility.download(bucket, serviceKey, file);
     }
 
@@ -153,7 +142,7 @@ public final class AWSS3StorageService implements StorageService {
             @NonNull File file,
             @NonNull ObjectMetadata metadata
     ) {
-        startServiceIfNotAlreadyStarted();
+        startTransferService();
         return transferUtility.upload(bucket, serviceKey, file, metadata);
     }
 
@@ -171,7 +160,7 @@ public final class AWSS3StorageService implements StorageService {
             @NonNull InputStream inputStream,
             @NonNull ObjectMetadata metadata
     ) throws IOException {
-        startServiceIfNotAlreadyStarted();
+        startTransferService();
         UploadOptions uploadOptions = UploadOptions.builder()
                                                     .bucket(bucket)
                                                     .objectMetadata(metadata)
@@ -187,7 +176,6 @@ public final class AWSS3StorageService implements StorageService {
      */
     @NonNull
     public List<StorageItem> listFiles(@NonNull String path, @NonNull String prefix) {
-        startServiceIfNotAlreadyStarted();
         ArrayList<StorageItem> itemList = new ArrayList<>();
         ListObjectsV2Request request =
                 new ListObjectsV2Request().withBucketName(this.bucket).withPrefix(path);
@@ -231,7 +219,6 @@ public final class AWSS3StorageService implements StorageService {
      * @param transfer an in-progress transfer
      */
     public void pauseTransfer(@NonNull TransferObserver transfer) {
-        startServiceIfNotAlreadyStarted();
         transferUtility.pause(transfer.getId());
     }
 
@@ -240,7 +227,7 @@ public final class AWSS3StorageService implements StorageService {
      * @param transfer A transfer to be resumed
      */
     public void resumeTransfer(@NonNull TransferObserver transfer) {
-        startServiceIfNotAlreadyStarted();
+        startTransferService();
         transferUtility.resume(transfer.getId());
     }
 
@@ -249,22 +236,11 @@ public final class AWSS3StorageService implements StorageService {
      * @param transfer A file transfer to cancel
      */
     public void cancelTransfer(@NonNull TransferObserver transfer) {
-        startServiceIfNotAlreadyStarted();
         transferUtility.cancel(transfer.getId());
     }
 
-    private void startServiceIfNotAlreadyStarted() {
-        if (!transferUtilityServiceStarted) {
-            // TODO: When a reset method is defined, stop service.
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-                Intent serviceIntent = new Intent(context, TransferService.class);
-                serviceIntent.putExtra(TransferService.INTENT_KEY_NOTIFICATION, createDefaultNotification());
-                context.startForegroundService(serviceIntent);
-            } else {
-                context.startService(new Intent(context, TransferService.class));
-            }
-            transferUtilityServiceStarted = true;
-        }
+    private void startTransferService() {
+        AmplifyTransferService.Companion.bind(context);
     }
 
     /**
@@ -274,32 +250,5 @@ public final class AWSS3StorageService implements StorageService {
     @NonNull
     public AmazonS3Client getClient() {
         return client;
-    }
-
-    private Notification createDefaultNotification() {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            createChannel();
-        }
-        int appIcon = R.drawable.amplify_storage_transfer_notification_icon;
-        return new NotificationCompat.Builder(
-            context,
-            context.getString(R.string.amplify_storage_notification_channel_id)
-        )
-            .setSmallIcon(appIcon)
-            .setContentTitle(context.getString(R.string.amplify_storage_notification_title))
-            .build();
-    }
-
-    @RequiresApi(api = Build.VERSION_CODES.O)
-    private void createChannel() {
-        NotificationManager notificationManager =
-            (NotificationManager) context.getSystemService(Context.NOTIFICATION_SERVICE);
-        notificationManager.createNotificationChannel(
-            new NotificationChannel(
-                context.getString(R.string.amplify_storage_notification_channel_id),
-                context.getString(R.string.amplify_storage_notification_channel_name),
-                NotificationManager.IMPORTANCE_DEFAULT
-            )
-        );
     }
 }