+++ testmodels/src/main/java/com/amplifyframework/testmodels/multiauth/OwnerUPPost.java
@@ -1,17 +1,16 @@
 package com.amplifyframework.testmodels.multiauth;
 
 
-import java.util.List;
 import java.util.UUID;
 import java.util.Objects;
 
+import androidx.annotation.NonNull;
 import androidx.core.util.ObjectsCompat;
 
 import com.amplifyframework.core.model.AuthStrategy;
 import com.amplifyframework.core.model.Model;
 import com.amplifyframework.core.model.ModelOperation;
 import com.amplifyframework.core.model.annotations.AuthRule;
-import com.amplifyframework.core.model.annotations.Index;
 import com.amplifyframework.core.model.annotations.ModelConfig;
 import com.amplifyframework.core.model.annotations.ModelField;
 import com.amplifyframework.core.model.query.predicate.QueryField;
@@ -29,6 +28,11 @@ public final class OwnerUPPost implements Model {
   private final @ModelField(targetType="ID", isRequired = true) String id;
   private final @ModelField(targetType="String", isRequired = true) String name;
   public String getId() {
+      return id;
+  }
+
+  @NonNull
+  public String resolveIdentifier() {
       return id;
   }