+++ testmodels/src/main/java/com/amplifyframework/testmodels/multiauth/PrivatePublicComboUPPost.java
@@ -5,6 +5,7 @@ import java.util.List;
 import java.util.UUID;
 import java.util.Objects;
 
+import androidx.annotation.NonNull;
 import androidx.core.util.ObjectsCompat;
 
 import com.amplifyframework.core.model.AuthStrategy;
@@ -106,7 +107,14 @@ public final class PrivatePublicComboUPPost implements Model {
     return new CopyOfBuilder(id,
       name);
   }
-  public interface NameStep {
+
+    @NonNull
+    @Override
+    public String resolveIdentifier() {
+        return id;
+    }
+
+    public interface NameStep {
     BuildStep name(String name);
   }